backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
rembg==2.0.32
pillow==9.5.0
python-multipart==0.0.6
